refactor(css): ajustar nomes de classes e funções
