+++ monitoramento-de-temperatura/interface/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
 
 <head>
     <meta charset="UTF-8">
@@ -9,7 +9,7 @@
     <title>Monitoramento de Temperatura</title>
 </head>
 
-<body>
+<body class="hide">
 
 
     <nav>
@@ -22,7 +22,7 @@
     </nav>
     
 
-    <div id="salasGrid" class="dashboard-grid"></div>
+    <div id="roomsGrid" class="room-grid"></div>
