+++ monitoramento-de-temperatura/interface/login.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
 
 <head>
     <meta charset="UTF-8">
@@ -8,7 +8,7 @@
     <title>Login</title>
 </head>
 
-<body class="body-page">
+<body class="default">
     <div class="container">
 
         <main>
@@ -16,14 +16,14 @@
                 <a href=""><img src="src/img/logo1.png" alt="logo do hospital sao camilo"></a>
                 <h2>Login</h2>
                 <form id="loginForm">
-                    <label for="usuario">Usuário </label><input type="text" name="usuario" id="usuario" required /><br>
+                    <label for="usuario">Usuário </label><input type="text" name="usuario" id="email" required /><br>
                     <label for="senha">Senha </label><input type="password" name="senha" id="senha" required /><br>
                     <button type="submit" class="btn">Entrar</button>
                 </form>
             </div>
         </main>
     </div>
-<script src="src/js/auth.js"></script>
+<script type="module" src="src/js/auth.js"></script>
 </body>
 
 </html>
