+++ monitoramento-de-temperatura/interface/relatorio.html
@@ -9,7 +9,7 @@
     <title>Monitoramento de Temperatura</title>
 </head>
 
-<body class="body-page">
+<body class="default">
 
     <nav>
         <a href="index.html"><img src="src/img/logo2.png"></a>
