+++ monitoramento-de-temperatura/interface/src/js/api.js
@@ -1,15 +1,17 @@
 
+//--------------------TEMPERATURAS----------------
 //buscar temperatura de todas as salas
 export async function buscarTemperaturas() {
   const token = localStorage.getItem("token");
   const response = await fetch("http://localhost:3000/api/temperatures",
-    {headers: {
-       "Content-Type": "application/json",
-      "Authorization": `Bearer ${token}`   
-    }
-  });
+    {
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+      }
+    });
 
-  if (!response){
+  if (!response) {
     throw new Error(`Erro na requisição: ${response.status}`);
   }
   const dados = await response.json();
@@ -36,17 +38,19 @@ export async function TempIntervalo(start, end) {
 }
 
 
+//---------------SALAS-------------------------
 //buscar temperatura de uma sala especifica
-export async function buscarTemperaturaSala(salaId)  {
-    const token = localStorage.getItem("token");
-    const response = await fetch(`http://localhost:3000/api/room/${salaId}/temperatures`,
-    {headers: {
-       "Content-Type": "application/json",
-      "Authorization": `Bearer ${token}`   
-    }
-  });
+export async function buscarTemperaturaSala(salaId) {
+  const token = localStorage.getItem("token");
+  const response = await fetch(`http://localhost:3000/api/room/${salaId}/temperatures`,
+    {
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+      }
+    });
 
-  if (!response){
+  if (!response) {
     throw new Error("Erro ao buscar temperaturas");
   }
   const dados = await response.json();
@@ -54,28 +58,49 @@ export async function buscarTemperaturaSala(salaId)  {
 }
 
 //buscar intervalo registrado de temperatura de uma sala especifica
-export async function roomTempInterval(salaId, start, end){
+export async function roomTempInterval(salaId, start, end) {
   const token = localStorage.getItem("token");
   const response = await fetch(`http://localhost:3000/api/room/${salaId}/temperatures/?start=${start}&end=${end}`,
-    {headers: {
-       "Content-Type": "application/json",
-      "Authorization": `Bearer ${token}`   
-    }
-  });
+    {
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+      }
+    });
 
-  if (!response){
+  if (!response) {
     throw new Error("Erro ao buscar temperaturas");
   }
   const dados = await response.json();
   return dados;
 }
 
+//Listar salas
+export async function roomsSearch() {
+  const token = localStorage.getItem("token");
+  const response = await fetch(`http://localhost:3000/api/rooms`,
+    {
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+      }
+    });
 
+  if (!response) {
+    throw new Error("Erro ao buscar temperaturas");
+  }
+  const dados = await response.json();
+  console.log(dados);
+  return dados;
+}
+
+
+//----------------------USUARIOS--------------------
 //Listar usuários do sistema
 export async function usersSearch() {
   const token = localStorage.getItem("token");
   const response = await fetch(`http://localhost:3000/api/users`,
-        {
+    {
       headers: {
         "Content-Type": "application/json",
         "Authorization": `Bearer ${token}`
@@ -90,22 +115,37 @@ export async function usersSearch() {
   return dados;
 }
 
+//Login de usuario
+export async function userLogin(email, senha) {
+  
+    const body = {
+        email: email,
+        password: senha
+    };
 
-//Listar salas
-export async function roomsSearch() {
-  const token = localStorage.getItem("token");
-  const response = await fetch(`http://localhost:3000/api/rooms`,
-        {
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${token}`
-      }
-    });
+    try {
+        // Faz a requisição de login
+        const response = await fetch("http://localhost:3000/api/login", {
+            method: "POST",
+            headers: {
+                "Content-Type": "application/json",
+            },
+            body: JSON.stringify(body),
+        });
 
-  if (!response) {
-    throw new Error("Erro ao buscar temperaturas");
+        const data = await response.json();
+        console.log(data);
+
+        if (response.ok) {
+            localStorage.setItem("token", data.token);
+            window.location.href = "index.html";
+
+        } else {
+            alert(data.message || "Erro no login.");
+        }
+    } catch (error) {
+        console.error("Erro ao fazer login:", error);
+        alert("Erro ao tentar fazer login. Tente novamente.");
+        console.log(data.token);
+    }
   }
-  const dados = await response.json();
-  console.log(dados);
-  return dados;
-}
