+++ monitoramento-de-temperatura/interface/src/js/components/charts.js
