+++ monitoramento-de-temperatura/interface/src/js/components/gauge.js
