+++ monitoramento-de-temperatura/interface/src/js/room.js
@@ -3,8 +3,8 @@ const params = new URLSearchParams(window.location.search);
 const salaAtual = params.get("sala");
 
 import { buscarTemperaturaSala, roomTempInterval } from './api.js';
-import { createGauge } from './gauge.js';
-import { tempChart, umidChart } from './charts.js';
+import { createGauge } from './components/gauge.js';
+import { tempChart, umidChart } from './components/charts.js';
 
 
 //carrega e atualiza a temperatura da sala a partir do id
@@ -26,32 +26,23 @@ async function roomUpdate() {
 
     const ultimaLeitura = dados[dados.length - 1];
 
-    const grid = document.getElementById("salaGrid");
+    const grid = document.getElementById("roomGrid");
     grid.innerHTML = "";
 
     const tempGaugeId = `gauge-temp-${salaAtual}`;
     const humGaugeId = `gauge-hum-${salaAtual}`;
 
     grid.innerHTML = `
-
-      <div class="temp">
-        <div id="${tempGaugeId}" style="width:320px; height:240px;"></div>
-        <h4>Temperatura</h4>
-      </div>
-      <div class="umid">
-        <div id="${humGaugeId}" style="width:320px; height:240px;"></div>
-        <h4>Umidade</h4>
-      </div>
-      </div>
- 
-    
-    
-`;
-
-
-
-
-
+              <div class="temp">
+                <div id="${tempGaugeId}" style="width:320px; height:240px;"></div>
+                <h4>Temperatura</h4>
+              </div>
+              <div class="umid">
+                <div id="${humGaugeId}" style="width:320px; height:240px;"></div>
+                <h4>Umidade</h4>
+              </div>
+          </div>    
+        `;
     createGauge(tempGaugeId, humGaugeId, ultimaLeitura);
 
   } catch (error) {
@@ -67,77 +58,6 @@ async function roomUpdate() {
 
 roomUpdate();
 setInterval(roomUpdate, 60000); // atualiza de 1 em 1 minuto
-
-
-
-
-function calcularIntervalo() {
-  const agora = new Date();
-  const start = new Date(agora);
-  start.setHours(agora.getHours() - 24); // Subtrai 24 horas para obter o "start" de 24 horas atrás
-
-  // Formatar as datas no formato ISO 8601 para passar na URL
-  const startISO = start.toISOString();
-  const endISO = agora.toISOString(); // A data de "agora" será o "end"
-
-  return { start: startISO, end: endISO };
-}
-
-
-// Função que busca e exibe a temperatura das últimas 24 horas
-async function tempIntervalRoom() {
-
-  const token = localStorage.getItem("token");
-
-  if (!token) {
-    alert("Sessão expirou, Faça login novamente");
-    window.location.href = "login.html";
-    return;
-  }
-  try {
-    const { start, end } = calcularIntervalo(); // Calcula o intervalo de 24 horas
-
-    const dados = await roomTempInterval(salaAtual, start, end);
-    console.log(dados);
-
-    if (!dados.length) {
-      console.error("Nenhum dado encontrado.");
-      return;
-    }
-
-    // Renderiza grid
-    const grid = document.getElementById("list");
-    grid.innerHTML = "";
-
-    dados.forEach(sala => {
-      const list = document.createElement("div");
-      list.className = "temp-sala";
-
-      list.innerHTML = `
-    <h3>${sala.room.name.toUpperCase()}</h3>
-    <h3>Temperatura: ${sala.temperature}</h3>
-    <h3>Umidade: ${sala.humidity}</h3>    
-    <p><em>Última atualização: ${new Date(sala.timestamp).toLocaleString("pt-BR")}</em></p>
-    <p>-----------------------------------------------------</p>
-  `;
-
-      grid.appendChild(list);
-    });
-
-  } catch (error) {
-    console.error("Erro ao carregar salas:", error);
-
-    if (error.message.includes("401") || error.message.includes("403")) {
-      alert("Acesso não autorizado. Faça login novamente.");
-      localStorage.removeItem("token");
-      window.location.href = "login.html";
-    }
-  }
-}
-
-// Chama a função intervalo
-tempIntervalRoom();
-setInterval(tempIntervalRoom, 60000); // Atualiza a cada 1 minuto
 
 
 
@@ -185,7 +105,27 @@ async function atualizarGraficos() {
   }
 }
 
+function calcularIntervalo() {
+  const agora = new Date();
+  const start = new Date(agora);
+  start.setHours(agora.getHours() - 24); // Subtrai 24 horas para obter o "start" de 24 horas atrás
+
+  // Formatar as datas no formato ISO 8601 para passar na URL
+  const startISO = start.toISOString();
+  const endISO = agora.toISOString(); // A data de "agora" será o "end"
+
+  return { start: startISO, end: endISO };
+}
+
+
 // Inicializa e atualiza a cada minuto
 atualizarGraficos();
 setInterval(atualizarGraficos, 60000);
+
+
+//eventListener de logout
+document.getElementById("logoutBtn").addEventListener("click", async function () {
+  localStorage.removeItem("token");
+  window.location.href = "login.html";
+});
